requirements.txt
yara-python==4.1.0
requests==2.25.1
python-dotenv==0.15.0
alive_progress